final/ticket-service/src/main/java/com/epam/training/ticketservice/models/Movie.java
package com.epam.training.ticketservice.models;

import javax.persistence.Entity;
import javax.persistence.Id;

@Entity
public class Movie {

    @Id
    private String movieName;

    private String type;

    private int duration;

    public Movie() {
    }

    public Movie(String movieName, String type, int duration) {
        this.movieName = movieName;
        this.type = type;
        this.duration = duration;
    }

    public String getMovieName() {
        return movieName;
    }

    public void setMovieName(String movieName) {
        this.movieName = movieName;
    }

    public String getType() {
        return type;
    }

    public void setType(String type) {
        this.type = type;
    }

    public int getDuration() {
        return duration;
    }

    public void setDuration(int duration) {
        this.duration = duration;
    }
}
